fix(android): hexLit byte handling for values > 127
Code review fix: use b.toInt() instead of b.toByte() in hexLit to
avoid sign-extension of UByte values > 127 when formatting hex.
Currently safe (BCD date values are < 128), but prevents future
regression if hexLit is used on larger values.

+++ android/app/src/main/java/com/pulseloop/ring/ColmiDecoder.kt
@@ -127,7 +127,7 @@ object ColmiDecoder {
         val marker = v[1].toInt()
         if (marker == 0xFF || marker == 0xF0 || v.size < 13) return emptyList()
 
-        fun hexLit(b: UByte): Int = String.format("%02x", b.toByte()).toInt()
+        fun hexLit(b: UByte): Int = String.format("%02x", b.toInt()).toInt()
         val year = 2000 + hexLit(v[1])
         val month = hexLit(v[2])
         val day = hexLit(v[3])
